Extract shardProwConfig and refactor it to re-use in tide shard branch cut tool
